Refactoring and renaming clickinput helpers
